Cars version 2
Browser mini-game version 2 in OOP with methods of Constructor in JavaScript

You don't need to change arguments of Objects because it will change automatically. 
Speed each of car be change in random in every 30 millisecond.
A gameplay background has also been added.
And now you can race on Tokyo Highway.
All comments in code did in the Russian language.